[libromdata] SNES: Support some more ROM mapping modes.
The BS-X title "SD Gundam G Next - Unit Map Collection" uses mode 0x23,
which is defined as LoROM+SA-1. This mode is generally not found in
regular SNES titles, but the BS-X header doesn't have a separate hardware
byte, so it's combined here.

(NOTE: There *might* be some SNES ROMs with this mode...)

This BS-X ROM is now detected properly.