[libromdata] SNES: Faster popcount() implementation.
This is used for compilers that don't have __builtin_popcount(),
e.g. MSVC. Note that MSVC does have __popcnt(), but that intrinsic always
emits the POPCNT instruction, so it requires manual cpuid checks. Since
we're only using popcount() for one thing, it isn't worth adding a
dispatch function for the POPCNT instruction.

Based on the version in gcc:
- https://gcc.gnu.org/bugzilla/show_bug.cgi?id=36041
- https://gcc.gnu.org/bugzilla/attachment.cgi?id=15529
- https://gcc.gnu.org/viewcvs/gcc?view=revision&revision=200506