[libromdata] GameCube: Calculate the size of Wii partitions.
The partitions are sorted by starting address, then the sizes are
calculated, similar to how rvthtool does it. This is needed for
unencrypted images from RVT-H systems, since the partition header
usually has the partition data size set to 0.

For other images (retail and RVT-R), the partition data size is valid,
so the calculated size isn't needed. It's done anyway just in case.

WiiPartition: Added a partition_size parameter to the constructor.
Note that this has the same name as a protected variable, so the
constructor has to use `this->partition_size` to access its own
variable instead of the constructor parameter.