Added checks for various headers and executables.
- libromdata: Check for `gl.h`. This is required by KhronosKTX.
  FIXME: This check isn't working with MSVC, even though the file is
  definitely present in the Windows SDK...

- kde5, libunixcommon: Check for `qtpaths`. This is required in order to
  determine the correct plugin directory for installation. Without it,
  we end up installing to /usr/plugins/, which is wrong.

This fixes issue #126. (Fedora: rp-stub has the wrong path for KDE5)
It was in fact related to issue #69 (QT_PLUGIN_INSTALL_DIR path problem),
with the problem being that the missing `qtpaths` executable was silently
ignored instead of failing with an error message.

In addition to failing on error, the error messages indicate which
packages need to be installed for Debian/Ubuntu and Fedora.