[win32] DllMain.cpp: Register RP_ShellPropSheetExt for "*" instead of each file extension.
This works around an issue where the properties tab doesn't show up if
Visual Studio 2017 is installed, since VS2017 sets up its own ProgID,
and Windows ignores the property sheet handlers for other ProgIDs.

Registering as "*" works, but it means the handler will be called for
all files, not just the ones we registered for. This is the same
behavior as the KDE frontend, though, so it should be fine.

This fixes issue #128. ([win32] Properties tab doesn't show up for .dds
if Visual Studio 2017 is installed)