[libromdata] DirectDrawSurface: Consolidated DDS format checking.
New function updatePixelFormat() replaces getPixelFormat() and handles:
- Standard DDS: Uncompressed textures
- Standard DDS: FourCC compressed textures
- DX10 DDS: Compressed and uncompressed

If a texture is uncompressed, the matching pixel format is set in
pxf_uncomp, along with bytespp for bytes per pixel.

If a texture is compressed, or the DX10 format is unrecognzied, its
format is set in dxgi_format using a DXGI_FORMAT value.

Note that DXGI_FORMAT doesn't directly support DXT2 or DXT4, so for
those cases, we also set is_premult to indicate that the texture has
premultiplied alpha.

Added several (untested) DX10 uncompressed RGB(A) formats.