[libromdata] DirectDrawSurface: Use DDS_DXT10_MISC_FLAGS2 to indicate alpha mode.
The DXT10 header has a field for alpha mode, which can allow for, among
other things, opaque DXT1, as well as DXT2 and DXT4. These normally
aren't selectable, since the DXGI format values only have BC1, BC2, and
BC3.

Make use of the alpha mode field to select between DXT2/DXT3 for BC2
and DXT4/DXT5 for BC5. Also use it to determine if DXT1 should have
1-bit alpha or if it should be opaque.

For non-DX10 images, set dxgi_alpha to the correct value based on the
FourCC or alpha mask.

TODO: For Luminance textures, set to standard alpha if it's Luma+Alpha?