[librpbase] formatFileSize(): Use C locale functions instead of C++.
The C++ locale functions only return `char` or `wchar_t`, which can
be a problem for non-ASCII characters. The C locale functions return
a char*, which fixes the problem for UTF-8.

On Windows, we always use localeconv() because nl_langinfo() isn't
available in MSVCRT. The MSVCRT version has both ANSI and Unicode
fields, so the Unicode field is converted from UTF-16 to UTF-8.

On other systems, we use nl_langinfo() if it's available. If it isn't,
we use localeconv().

References:
- http://utf8everywhere.org/ (section 9)
- https://www.gnu.org/software/libc/manual/html_node/The-Lame-Way-to-Locale-Data.html#The-Lame-Way-to-Locale-Data
- https://www.gnu.org/software/libc/manual/html_node/The-Elegant-and-Fast-Way.html
- https://msdn.microsoft.com/en-us/library/8658zdx3.aspx