[win32] Split the en_US resources out of resource.rc.
Each language will have its own resource file. We can't use gettext for
the dialog resources because, unlike Qt and GTK+, Windows doesn't have
built-in layout management, so the dialogs have to be adjusted for each
language.

Added `#pragma code_page(65001)` to the .rc files to indicate that
they're encoded using UTF-8.

VS_VERSION_INFO: Updated the copyright date. Keeping this in the common
resource.rc file for now because it's unlikely to be localized, though
I'm leaving the language code set to en_US.

IDD_CONFIG_IMAGETYPES: Changed the language to en_US. This was previously
set to "neutral". It's now located in the rp-en_US.rc file.

Moved the .rc files and resource.h into res/. Note that paths are still
relative to the project root, so we still have to specify "res\\".