[win32] ConfigDialog: Added translations for various strings.
Moved the property sheet title initialization to WM_SHOWWINDOW. We can't
store a pointer to wstring::c_str() in the PROPSHEETHEADER because the
pointer would go out of scope, and we don't want to store the wstring in
the private class because that's a waste of memory, so setting it in
WM_SHOWWINDOW makes sense.

Minor optimization for the "Reset" button.