[win32] RP_ExtractIcon_Register.cpp: Delete "OpenWithProgids/rom-properties" if it's present.
This was used before for icon associations, but now it seems to break the
icon handler on Windows 7, probably because we're not registering the
icon handler in the rom-properties ProgID anymore due to the
implementation of icon handler fallbacks. (On Windows XP, it didn't cause
any issues.)

NOTE: It seems the code that used "OpenWithProgids" wasn't actually
committed. I was probably testing some stuff to try working around the
DDS property tab issues with MSVC installed. Either way, we should ensure
that this gets removed if it's present.