[libwin32common] RegKey::RegisterApprovedExtension(): Create "Shell Extensions\\Approved" if it doesn't exist.
This key doesn't exist on ReactOS 0.4.8, so we should create it instead
of failing.

This fixes component registration on ReactOS. Note that ReactOS's
Explorer doesn't support icon or thumbnail previews yet, but it does
support property sheet handlers, and RP_ShellPropSheetExt appears to
work as expected.

NOTE: Not adding an entry to CHANGES.md because no one uses ReactOS
in production.