Rewrote DelayLoadHelper in C.
It didn't use any C++-specific functionality, so this mostly consisted
of reordering things to compile with MSVC 2010.

DelayLoadHelper.c:
- Reworked to compile with MSVC 2010. (C89)
- Check _MSC_FULL_VER for VS2015 Update 3 instead of _MSVC_LANG, since
  _MSVC_LANG is only defined in C++ mode.
- Moved the architecture-specific stuff out of rp_loadLibrary().

MSVC 2010 (32-bit) build of DelayLoadHelper.obj (in win32/):

|     Build    |   Before   |   After   |   Diff   |
|-------------:+-----------:+----------:+---------:|
|   Debug .obj |     31,865 |    29,216 |   -2,649 |
| Release .obj |     21,940 |    21,874 |      -66 |
|   Debug .dll |  2,399,774 | 2,400,256 |     +482 |
| Release .dll |    718,848 |   719,360 |     +512 |

The .obj file sizes decreased, but the .dll file sizes increased,
possibly because we have to build stdafx.c now due to the use of
C code in the win32 and rpcli projects.

[win32] stdafx.h: comdef.h only works with C++ (MSVC 2010), so only
include it if building as C++.