[librpbase] rp_image::squared(): Fixed heap corruption when squaring a 256x512 VTF3 image.
Fixed several issues that occurred when squaring an image that's taller
than it is wide, resulting in heap corruption:

- Copy the source row bytes, not the destination row bytes. Destination
  row bytes is always larger than the source.
- Calculate the correct dest_blanking value. Previously, it would get
  the difference between sq_img's stride and row bytes, which is usually
  zero if there's no padding.
- Main loop: Check for `y >= 0` and start with `(height-2)` to make it
  more obvious that we're skipping the last line.

Note that rp_image::squared() is currently only used by the Win32
thumbnailer. This bug may affect anything that has taller-than-wide
images, e.g. amiibo and textures.

This should either be backported to 1.2 and released as 1.2.1, or we
need to finish up 1.3 quickly.

TODO: Add a unit test for rp_image::squared().