[libromdata] KeyStoreUI: Check file->isOpen(), not if file is nullptr.
file can't be nullptr due to C++ semantics.

This fixes an issue where the UI indicated the key file was invalid in
cases where the actual issue was it couldn't be read, e.g. the file is
chmod 0000. It now shows that an error occurred while opening the file.

TODO: Return the POSIX error code.