[libromdata] GameCom: Only adjust the icon offset if the file is 0.8 MB or 1.8 MB.
It seems the header being at 0 is actually allowed. The 0.8 MB ones
truncate the first 256 KB on ROMs where the header is at 0x40000, which
is why the adjustment is needed.

This fixes the following icons:
- Henry
- Indy 500
- Jeopardy!
- Wheel of Fortune
- Wheel of Fortune 2

Still broken:
- Fighters MegaMix [b1]
- Game.com Internet
- Tiger Web Link

NOTE: Game.com Internet and Tiger Web Link don't seem to actually have
icons in the normal format. These cartridges might override the BIOS in
some way.