[libromdata] GameCom: Fixed icons whose Y positions are not a multiple of 4.
Reworked the blitting code such that X is the outer loop, since we have
to rotate the image.

If the icon is aligned to a Y multiple of 4, it's blitted directly to
the rp_image as before. If not, it's blitted to a temporary buffer,
which is 68 lines tall. The temporary buffer is then copies to the
rp_image.

This fixes:
- The Lost World
- Monopoly
- Quiz Wiz - Cyber Trivia
- Wheel of Fortune
- Wheel of Fortune 2

Batman and Robin's icon is actually misaligned in the ROM, so there's
nothing to be fixed for that here.

As of this commit, all valid game.com icons should be working now.