[libromdata] TImageTypesConfig: Removed Dreamcast.
I was going to add game.com, but I decided not to because it only has a
single internal image and no external images. Dreamcast is currently the
same, so I decided to remove it from here to reduce the size of the
dialog.

These can be added later if external images are added.

This effectively reverts commit 69acffcd6cf4554acef478f11193ae6b705ae085.
([libromdata] TImageTypesConfig: Added Dreamcast.)