Enabled delay-load for libgnuintl-8.dll.
libintl.h: Disabled libgnuintl wrappers for printf() functions.
We don't need to use the GNU versions for anything, and it causes problems
if libgnuintl-8.dll is unavailable, since we can't use printf() to print
an error message in that case.
- Redefine snprintf() as _snprintf() on older versions of MSVC.
  See c99-compat.msvcrt.h.

DelayLoadHelper.cmake: Delay-load libgnuintl-8.dll.

DelayLoadHelper.h: Added a 1-argument macro.

[win32] DllMain.cpp, DllGetClassObject() Fail if libgnuintl-8.dll can't
be loaded.

[rpcli] rpcli.cpp: Print an error message if libgnuintl-8.dll can't be
loaded.
- TODO: Print the program information.

[win32] ConfigDialog.cpp: Show an error message if libgnuintl-8.dll
can't be loaded.
- TODO: Move this to rp-config.c? Doing that will result in no error
  message if rundll32.exe is used, though...

This should be the last thing needed for issue #122:
[win32] Build and link in GNU gettext or equivalent