[win32] AboutTab: Preparations for RichEdit 4.1 "friendly links".
This allows us to set both the title and link content when using
RichEdit 4.1.

Load both RichEd20.dll and MsftEdit.dll. If MsftEdit.dll is available,
it will be used for RichEdit 4.1; otherwise, RichEdit 2.0 will be used.

New function rtfFriendlyLink() that creates a friendly link if
supported, and falls back to just the title if not.
- Reference: https://blogs.msdn.microsoft.com/murrays/2009/09/24/richedit-friendly-name-hyperlinks/

TODO: Actually use RichEdit 4.1 if it's available.

[kde] AboutTab: Use the same localized strings as Win32 for the program
license. Remove my name from the hyperlinked portion of the email
support link on the "Support" tab.