[win32] CacheTab: Dynamically add/remove/update devices from the ListView.
ConfigDialog: Forward WM_DEVICECHANGE messages to CacheTab. When a
device is changed, it's queried for new information, then added,
removed, or updated in the ListView as needed.

NOTE: The update is done 1 second after receiving the message in order to
ensure that the drive has stabilized if we're inserting/ejecting a CD.
We might be able to reduce this value later, or maybe just keep retrying
until SHGetFileInfo() returns success.

TODO:
- Optimize by storing a map of drive letters to ListView indexes
  somewhere.
- Always sort by drive letter?