[gtk] CairoImageConv: Premultiply images and/or palettes before converting to cairo_surface_t.
For CI8, we only have to premultiply the palette. This is done using a
new rp_image function, rp_image::premultiply_pixel(), which is a wrapper
around un-premultiply.cpp's premultiply_pixel_inl() static inline
function. (We're not exporting the static inline function because that
reduces the performance of rp_image::premultiply().)

For ARGB32, we dup() the image, then premultiply it.
- TODO: Combined dup() and premultiply() function? Though it'd only be
  used in one place right now...