[kde] Initial KFileMetaData extractor plugin. (KDE5 only!)
Note that Qt's plugin system only allows exporting a single plugin
interface per .so file. ThumbCreator doesn't use the Qt plugin system,
but KPropertiesDialogPlugin does. Hence, for KFileMetaData, we're using
a forwarder library that exports the plugin, then creates the actual
plugin by calling into rom-properties-kde5.

Currently, only ADX files ("audio/x-adx") are supported, and a fixed
value (1234 seconds; 20:34) is returned. The next step is adding metadata
functions to the RomData subclasses.

NOTE: On Dolphin 5.46.0, file metadata doesn't show up unless system-wide
indexing is disabled. I'm guessing that if indexing is enabled, it only
reads from the cache, and these values haven't been cached yet because
the plugin is new; whereas if it's disabled, it fetches the properties
immediately.

FIXME: Ubuntu 16.04 doesn't have the KDE4 version of KFileMetaData.
I'll need to install an older version in order to ensure the plugin
works there. KFileMetaData is disabled on KDE4 for now.

NOTE: The RpExtractorPlugin and RpExtractorPluginForwarder files are
located in kde5/. This is because Qt's moc doesn't supprt preprocessor
directives like `#ifdef`, so we'll need separate versions for each.