[libromdata] RomDataFactory: Store the results of supportedFileExtensions() and supportedMimeTypes().
There's no point in having to gather the information multiple times
if the caller requests it multiple times.

The main functions now call pthread_once() for internal functions.
A const reference is returned instead of the actual vector.

NOTE: Since the vector is static and not a pointer, it should be
automatically deleted when the shared library is unloaded.

Updated callers to use const references.