[libromdata] Dreamcast, SegaSaturn: If the file is too small, assume it's a boot sector.
IP.bin is 32 KB. A Dreamcast or Saturn disc image smaller than 64 KB is
highly unlikely, so if the file is <= 64 KB, assume it's a boot sector.

NOTE: This is only checked if the file has 2048-byte sectors, since the
raw boot sector wouldn't have the extra information from a 2352-byte
sector.

[librpbase] RomData: Added FTYPE_BOOT_SECTOR.