[win32] DllMain: Register a ProgID again.
We're not setting the default ProgID in file extensions, but we *are*
adding rom-properties to OpenWithProgids.

This is needed for proper handling of RP_PropertyStore.

FIXME: .dds icons don't preview on Win7 if MSVC 2017 is installed.
Thumbnails work fine, though. Note that this was apparently not working
in v1.3, either... (It works on XP and Win10; might be a Win7 bug.)

RP_ExtractIcon_Register.cpp: Removed the "OpenWithProgids" code.
This didn't work in all cases because it only runs if the file extension
has thumbnails, but we registered OpenWithProgids for all file extensions
previously in order to handle property sheets.