[win32] RP_PropertyStore: Initial metadata class for Windows.
This class currently exports four hard-coded properties, but only two
of them are defined to be visible in "PreviewDetails". "FullDetails",
which is shown as a tab in file properties, will be added later.

RP_PropertyStore is registered in the ProgID instead of as part of
each file extension. This required adding HKLM (and/or user root)
to DllMain's RegisterFileType() and related functions.

TODO:
- Standard properties and ordering.
- Get metadata from the RomData object.
- RomData doesn't provide "Dimensions", but this can be created
  from Width and Height, and possibly Depth.