[libromdata] WiiWAD: Display a few fields.
These fields are taken from the unencrypted TMD header.

CBCReader is now initialized, but it isn't used for anything yet.
The encryption key is loaded from KeyManager, and it's currently
hard-coded to "rvl-common".

The encryption key status is also displayed if the key could not
be loaded for some reason.

TODO:
- Determine the correct key by checking the TMD issuer and key index.
- Add key verification.