[libromdata] WiiWAD: Display title version and region code.
Title version is displayed as both major.minor and vNUMBER.

Region code has a special case for IOS (region-free) and system menu
(based on the title version). For everything else, the fourth character
from the ID4 is used.

TODO: Is there a real way to get this from the WAD contents?