[librpbase] RomMetaData: KDE's ReleaseDate and TrackNumber are unsigned.
Based on baloo_filemetadata_temp_extractor output from a .flac file.

Added unsigned integers to RomMetaData. They're technically handled the
same as signed, so it really depends on how it's added.

KDE: Use the correct type so the QVariant is set up correctly.

Windows: The type is set by COM, so it was already "correct". Use the
correct union field based on UIx vs. Ix.