[librpbase] RpFile_stdio.cpp: Added transparent gzip decompression.
RpFile.hpp: New modes FM_GZIP_DECOMPRESS and FM_OPEN_READ_GZ.

If FM_OPEN_READ_GZ is specified, the file is checked for a gzip header.
If found, it will be reopened and processed using zlib.

Transparent gzip decompression opens a gzFile. When dup()'ing the
file handle, we have to reopen the original file handle instead of
dup()'ing it, since the gzFile is associated with the original
handle. dup()'ing it would cause confusion.

Use FM_OPEN_READ_GZ in the rpcli, KDE, and GTK+ frontends.

[win32] Use FM_OPEN_READ_GZ in RP_ExtractIcon, RP_ExtractImage,
and RP_ShellPropSheetExt.

TODO: RP_ThumbnailProvider and RP_PropertyStore use IStream, so we'll
need to manually implement an IStream gzip wrapper.