Revert "[librpbase] RpPngWriter: Removed a redundant `file = nullptr;` assignment."
This reverts commit 2a3416c666c29d361a30d5c3518a353b0d9ade03.

Turns out it wasn't quite redundant. Not having this causes `rpcli` to
crash when extracting an animated icon.

NOTE: This means the `-a` option is broken in rom-properties v1.3.x.
I might cherry-pick this revert in the release/1.3 branch if I have to
release v1.3.4 for a major issue later.