[libromdata] WiiWIBN: Initial WIBN (Wii save game banner) parser.
This can be used to preview unencrypted save game banners, and will later
be used with CBCReader to preview encrypted save games.

NOTE: Some Wii game saves have animated icons where the first frame has
a delay value of 0, which normally means "end". This should be ignored
and the icon should be displayed as-is.

TODO: Animated icon timing seems to be 50% of what it should be. We might
need to adjust GCN icons, too.