[libi18n] i18n.c: Wrap the file in #ifdef ENABLE_NLS.
We don't want anything compiled in here if NLS is disabled because
rp_i18n_init() is defined as a dummy macro in that case.

We can't simply disable compilation of the file because CMake needs at
least one file in order to build the static library, so wrapping it in
an #ifdef will work instead.