[librpbase] RpFile_stdio.cpp: delete d_ptr; on deletion.
Otherwise, the files are never closed.

This never seemed to be a serious problem on Linux, since thumbnails are
run out of process and files aren't locked when they're opened, but the
property pages are run *in* process, so it's entirely possible that the
file handle leaks could cause an EMFILE (Too many open files) error.

This happened when running `rpcli` on N64 ROM headers, which failed on
the 1,023rd file. (Limit of 1,024 open files?)