[libromdata] SegaSaturn: Added metadata properties.
Same properties as Dreamcast.

Split code from loadFields() into getPublisher() and parseDiscNumber(),
similar to Dreamcast. We can't share parseDiscNumber() between the two
classes, but getPublisher() might be able to be shared between them,
and maybe MegaDrive later.

SegaSaturnPrivate::getPublisher(): Check that `*endptr == ' '`. This was
present in the Dreamcast version.

DreamcastPrivate::getPublisher(): Check `t_code != 0`. This was present
in the SegaSaturn version.