[libromdata] GameCube: Display title ID and access rights for Wii games.
Title ID low is usually the same as the game ID, except for some
diagnostics discs.

Access rights are usually both blank, but diagnostics discs and
Homebrew Channel have AHBPROT set. The old DVDX channel has the
DVD Video flag set. Presumably, an official DVD player with that
flag set was also in the works, but it was never released.