[libromdata] SNDH: Skip over the string data, plus 1 for the NULL terminator, after reading it.
Otherwise, we end up searching for tags in the string data. We probably
won't find one, but if we do, that'd result in garbage text. Also, it
wastes time.

This fixes ripper tags for files that have e.g. "NOT RIPPED", since
previously the "RIPPED" string would be interpreted as another 'RIPP'
tag, causing it to be read as "ED".

FIXME: Files with non-ASCII titles aren't being handled correctly,
since we're skipping the number of UTF-8 bytes, not the original
number of bytes.