[libromdata] SNDH: Initial 16-bit tag parsing, starting with number of subtunes.
readStrFromBuffer() now takes a pointer to `const char*`, which
allows it to increment the pointer past the NULL terminator directly
instead of requiring the caller to do so.

New function readAsciiNumberFromBuffer(), which is similar, but uses
strtoul() to convert the ASCII number to an unsigned int. This is
used for 'YEAR' and '##'.