[googletest] Build as a DLL on Windows.
This significantly reduces the size of the test executables.

This can be done more easily on Windows than on other platforms because
Windows includes the program's directory as part of the DLL search path
by default.

NOTE: I was originally trying to fix a linkage issue with
g_linked_ptr_mutex that was showing up when building on MSVC 2010,
but that seems to have disappeared by itself...

References for the linkage issue:
- https://github.com/google/googletest/issues/292
- https://github.com/ornladios/ADIOS2/pull/477
- https://github.com/ornladios/ADIOS2/pull/477/commits/d9657bda86f714ff7b3c2f3d62f041808196ee39