[gtk] Use glib-mkenum to generate enumeration files.
It's not integrated into the build system due to issues with custom
commands and the fact that the gtk directory doesn't have a target
by itself. Instead, an mkenums.sh script is provided to update the
rp-gtk-enums.c and rp-gtk-enums.h files, which are both checked into
the repository.

Removed the relevant code from RomDataView.cpp and RomDataView.hpp,
since it's now in rp-gtk-enums.c and rp-gtk-enums.h.