[libromdata] WiiPartition: Split crypto method into sector size and encryption.
NASOS images are unencrypted, but they still use the 1K/31K split,
presumably to make it easier to convert back to the regular encrypted
image format.

This lets us specify unencrypted and 1K/31K for NASOS, unencrypted
and full 32K for RVT-H, and encrypted and 1K/31K for standard.

TODO:
- Test all image formats and make sure nothing broke.
- Show the key that would be used if it was a regular image.