[libromdata] GameCube: Improve "default" ID4 region checks.
gcnRegionToString(): Added an optional pointer-to-bool parameter. This
pointer will be set to `true` if the ID4 matches the default, and
`false` if it doesn't.

Note that only codes that specifically map to a country are considered
to not be default. Generic codes like 'X', 'Y', and 'Z' are handled as
if they're the default country.

loadFieldData(): Updated to handle this change.