[libromdata] Nintendo3DS: Remove the isSysTitle and isSigZero checks for "dangerous" permissions.
Homebrew titles seem to have a non-zero signature. I'm not sure exactly
what they're using for signatures, but that prevents the zero check
from working.

Show the overlay for system titles. We can't verify that it's a Nintendo
signature without implementing RSA code, and chances are users won't be
installing system titles randomly.