[librpbase] FileSystem_win32.cpp: Removed the GetFinalPathNameByHandleW() declaration.
We're using GetProcAddress() for this, and GetFinalPathNameByHandleW() is
declared in the Windows SDK as of Windows Vista. I probably added this
for testing purposes before increasing the minimum SDK requirements to
Windows Vista.

MSVC 2010 complained about it:
FileSystem_win32.cpp(60): warning C4273: 'GetFinalPathNameByHandleW' : inconsistent dll linkage

It turns out MSVC 2017 complained about it too, but I didn't notice it
before.