[librpbase] ImageDecoder_BC7.cpp: Added preliminary BC7 decoding.
BC7 doesn't use S3TC, so it isn't dependent on ENABLE_S3TC.

The decoder currently only determines the mode for each block and sets
a fixed color based on the mode setting. I encoded argb-reference.png
using LSDxt 4.0 and the resulting colors (modes) are:
- Cyan (Mode 3)
- Red (Mode 4)
- Magenta (Mode 5)
- Yellow (Mode 6)

With rgb-reference.png, it's basically the same, but with less magenta
and more cyan.