[librpbase] ImageDecoder_BC7: Fix Mode 4 index selection behavior.
If idxMode == 0, color data uses the 2-bit indexes, and alpha data uses
the 3-bit indexes. If idxMode == 1, it's swapped.

The code was handling color data correctly, but alpha data was using the
wrong indexes with the wrong bit size.

w5_rock805_nrm.dds now *looks* much closer to what MSVC 2017 shows, but
the color values are still off. (0,0) is RGB(100,135,255) on MSVC, but
rpcli's version is RGB(124,85,248). The B component can be explained by
color expansion issues, but not the R or G components...