[libromdata/tests] CMakeLists.txt: Copy BC7/, not BCn/.
I was going to put both BC5 and BC7 images into the same directory, but
then decided to change it to separate directories. BC5 isn't in yet,
and I forgot to change the CMakeLists.txt entry to BC7.

This should fix the travis-ci build.

I'm guessing it worked locally because I had it set to BC7 at one point,
but then some random patches and stuff reverted it.