[librpbase] ImageDecoder_BC7, rshift128(): Don't subtract 64 from `shamt`.
This was part of an attempt to support shift amounts larger than 64, but
that never happens here.

On 64-bit systems, the shift amount is modulo 64, so this has no effect.

On 32-bit systems, the shift amount is modulo 32, so the extra code gcc
and MSVC emits to handle 64-bit shifts gets confused and it does the
wrong thing.

This should fix the 32-bit Windows builds.

FIXME: The Mac OS X build on travis-ci is failing, possibly due to
something screwy with zlib.