[librpbase] ImageDecoder_BC7: Optimize P-bits by using 32-bit writes.
endpoints is now a union of u8 and u32. u8 has been expanded from
[6][3] to [6][4] to allow for 32-bit writes.

P-bit code: If the P-bit is 1, apply a mask of 0x01010101 << shift
amount to the u32 value. For Mode 1, this is hard-coded to 0x02020202.
For others, a shift is done at runtime.

common.h: New ALIGNED_VAR() macro to align a stack and/or static
variable. This is used for fromBC7()'s tileBuf[].

BC7 benchmarks: (100 iterations, 64-bit Gentoo Linux, release build, gcc-8.2.0-r6)

|        Filename        |  Before  |  After   |      Difference      |
|:-----------------------|---------:|---------:|---------------------:|
| w5_grass_200_abd_a.dds |    34 ms |    28 ms |      -6 ms (-19.35%) |
| w5_grass201_abd.dds    |   473 ms |   469 ms |      -4 ms (-0.849%) |
| w5_grass206_abd.dds    | 3,741 ms | 3,737 ms |      -4 ms (-0.107%) |
| w5_rock805_abd.dds     |   397 ms |   393 ms |      -4 ms (-1.013%) |
| w5_rock805_nrm.dds     |   495 ms |   492 ms |      -3 ms (-0.608%) |
| w5_rope801_prm.dds     |    14 ms |    14 ms |       0 ms ( 0.000%) |
| w5_sand504_abd_a.dds   |   475 ms |   478 ms |      +3 ms (+0.630%) |
| w5_wood503_prm.dds     |   485 ms |   481 ms |      -4 ms (-0.828%) |

Not a significant performance difference, but there is a pretty good
code size reduction:

Code size difference:
   text    data     bss     dec     hex filename
   4627       8       0    4635    121b ImageDecoder_BC7.cpp.o [before]
   4554       8       0    4562    11d2 ImageDecoder_BC7.cpp.o [after]
    -73       0       0     -73     -73 Difference