[librpbase] ImageDecoder_BC7: Removed SSSE3 code.
It didn't seem to help in various benchmarks besides the initial one,
which may have been caused by a bad test environment. AppVeyor's
MSVC 2013 environment was segfaulting in the BC7 SSSE3 tests, but only
in 32-bit debug builds. I was unable to reproduce this locally on
any of MSVC 2010, 2013, or 2017.

Maybe I'll try optimizing it again later...