[libromdata] DMG: Added metadata properties.
Split the title, game ID, and publisher code from loadFieldData() into
separate functions. Note that title and game ID are in a single function
because the size of the title field, and presence of game ID, vary
depending on whether or not it's a CGB ROM and if it's older or newer.