[librpbase] TextFuncs: Refactor the conversion functions into wrappers around cpN_to_utf8() and related.
The macro-ized functions introduced in v1.2 were a significant
improvement over the duplicated lines of code we previously used,
but it still took up a lot of space in the binary. Using the
cpN_to_utf8() and related functions lets us use a parameter in a
single instance of a function instead of duplicating it.

Code size reductions:

64-bit Gentoo Linux, gcc-8.2.0-r6, release build:
   text    data     bss     dec     hex filename
   6704       8       0    6712    1a38 TextFuncs_iconv.cpp.o [before]
   5031       8       0    5039    13af TextFuncs_iconv.cpp.o [after]
  -1673       0       0   -1673    -689 Difference

MSVC 2017 15.9.5, 64-bit, release build:
   text    data     bss     dec     hex filename
   8911       8       8    8927    22df TextFuncs_win32.obj [before]
   4062       8       8    4078     fee TextFuncs_win32.obj [after]
  -4849       0       0   -4849   -12f1 Difference