[librpbase] RpFile_win32.cpp: Cast all DWORDs to int64_t before multiplying.
MSVC 2017's code analysis tool says that not doing this could result in
a 32-bit multiply, which might truncate the value.

C26451: Arithmetic overflow: Using operator '*' on a 4 byte value and
then casting the result to a 8 byte value. Cast the value to the wider
type before calling operator '*' to avoid overflow (io.2).