[libromdata] XDBF: Initial Xbox 360 XDBF parser.
Xbox 360 executables (XEX) contain an XDBF resource section, which
has, among other things, the game name and icon.

This is a very rough preliminary version that manually parses the
string table to find the game name. The string table handling should
be split out into a separate function, as should the general resource
lookup function.

Note that XDBF resources use a very different format compared to
Windows resources. In particular, strings are encoded as UTF-8
instead of UTF-16, and icons are stored in PNG format.

A good amount of documentation was used from the Free60 project,
Xenia, and XexTools to implement this functionality.