[libromdata] Xbox360_XEX: Initial support for reading the XDBF section.
If the PE executable is encrypted, CBCReader is used to decrypt it on the
fly. Encryption is AES-128-CBC, with IV=0. The AES title key is encrypted
with a common key on retail, and all zeroes on debug.

FIXME: sonic06.xex is actually "compressed", but the compression seems
to merely remove zero padding between sections. Need to figure out how to
uncompress it. For now, I hard-coded the XDBF offsets, so only sonic06.xex
will thumbnail correctly.

CBCReader: Added support for reading unaligned blocks. This wasn't needed
for the Nintendo 3DS classes, which has alignment for everything, but for
XDBF, we do unaligned reads regularly. Aligned reads would be faster,
though, since we don't have to throw away unused decrypted blocks.

TODO:
- If the retail key fails, try the debug key.
- Handle compressed PE executables correctly.