[libromdata] BCSTM: Sample block count includes the last block.
We have to subtract 1 because the last block isn't necessarily a full
block.

This fixes an error where the length of the track appears to be slightly
longer than it is, which is especially noticeable if the track loops and
the loop end is set to the end of the track.

NOTE: If there's 0 samples, this will underflow, but that would be an
invalid file.