[libromdata] GdiReader::openIsoRomData(): Get an ISO object for the specified track.
Dreamcast: Call GdiReader::openIsoRomData() if we're reading a GDI image.
This lets us show the PVD information for GDI images.

NOTE: Since the ISO object takes a reference to the PartitionFile created
in GdiReader, GdiReader doesn't actually need to keep a pointer to it.
We should do this in other classes.