[libromdata] XboxDisc: Show the disc type.
This includes the XGD version and, for XGD2, the manufacturing wave.

The PVD version of isRomSupported() now has an optional pWave
parameter that, if specified, receives the wave number. (0 for error)

Added the XGD1 timestamp. I'm not sure if all XGD1 discs use the same
timestamp, but so far it looks like it. The tz_offset field in the PVD
looks wrong, but we're adjusting the timestamp to match the value it has.
(0x30, ASCII '0')