[libromdata] XboxDisc: New private function getConsoleType().
This first checks for the presence of default.xex vs. default.xbe.
If default.xex is found, the game is assumed to be for Xbox 360.

If neither is found, XGD2 and XGD3 are assumed to be for Xbox 360,
and XGD1 and extracted XDVDFS are assumed to be for the original Xbox.