[librpbase] RpFile: New function rereadDeviceSizeScsi().
This rereads the device size using SCSI commands. It's only useful on
Windows, since RpFile_win32 caches the device size.

Something seems to be off, though:
- Full XGD2 size: 7,834,892,288
- Size reported by READ CAPACITY(10): 7,572,881,408
- Difference: 262,010,880

This might not be a problem if the offsets don't change. Will need to
do more tests.

The XDVDFS partition still can't be read on Windows yet, since we're
still using Windows API instead of SCSI READ(10).