[libromdata] GameCube, WiiWAD: Added China and Taiwan region codes.
These weren't used in retail software, but some development software
with the China region code exists. I haven't seen Taiwan region yet,
but I assume it would be the next value.

We're assuming CHN and TWN uses Shift-JIS encoding. It might be something
else, like Big5 and/or GB2312, but I don't have anything to check that
at the moment.