[librpbase] PartitionFile: Initial infrastructure to cache small files on devices.
This adds an optional `cache` parameter that, if set to true, tells
PartitionFile to cache the entire file if it's smaller than 4 MB.

TODO: Rework this to remove the `cache` parameter and always cache
the file if it's smaller than 4 MB and being read from a block device.
This will reduce seeking when reading e.g. Xbox 360 discs using a
Kreon drive.