[libromdata] Xbox360_XEX: Properly parse the resource information header.
getXdbfResInfo() now loads the execution ID. If not found, it won't be
able to load the resources.

The title ID (which is found in the execution ID) is used to determine
which resource has the XDBF. Most games only have one resource, but some
newer games (including Gears of War 3) also have a "HashSec" resource,
which I assume is for extra security of some sort.

XEX2_Resource_Info: Removed the `size` field. This struct now only
contains a single resource information entry.