[libromdata] Xbox360_XEX: Display the minimum kernel version required for the XEX.
This corresponds to the dashboard version.

This is determined by checking all of the import libraries and taking
the highest version number in the "minimum required" field.

xbox360_xex_structs.h: Added an XEX2_Version_t struct with correct
bitfield ordering for both big-endian and little-endian, assuming
the value is in host-endian.

TODO: Show the XEX2 version numbers?